Add mutable.Signal.imap, revert changes on immutable.Signal
