+++ core/src/main/scala/fs2/async/immutable/Signal.scala
@@ -41,28 +41,15 @@ trait Signal[F[_], A] { self =>
    * Asynchronously get the current value of this `Signal`
    */
   def get: F[A]
-
-  /**
-   * Returns an alternate view of this `Signal` where its elements are of type [[B]],
-   * given a function from `A` to `B`.
-   */
-  def map[B](f: A => B)(implicit F: Functor[F]): Signal[F, B] =
-    new Signal[F, B] {
-      def discrete: Stream[F, B] = self.discrete.map(f)
-      def continuous: Stream[F, B] = self.continuous.map(f)
-      def changes: Stream[F, Unit] = self.changes
-      def get: F[B] = F.map(self.get)(f)
-    }
 }
-
 
 object Signal {
 
-  implicit class ImmutableSignalSyntax[F[_] : Async,A] (val self: Signal[F,A])  {
+  implicit class ImmutableSignalSyntax[F[_] : Async, A] (val self: Signal[F, A])  {
     /**
      * Converts this signal to signal of `B` by applying `f`
      */
-    def map[B](f: A => B):Signal[F,B] = new Signal[F,B] {
+    def map[B](f: A => B): Signal[F,B] = new Signal[F, B] {
       def continuous: Stream[F, B] = self.continuous.map(f)
       def changes: Stream[F, Unit] = self.discrete.through(pipe.changes(_ == _)).map(_ => ())
       def discrete: Stream[F, B] = self.discrete.map(f)
@@ -84,10 +71,5 @@ object Signal {
   def hold[F[_],A](initial: A, source:Stream[F,A])(implicit F: Async[F]): Stream[F,immutable.Signal[F,A]] =
     Stream.eval(fs2.async.signalOf[F,A](initial)) flatMap { sig =>
       Stream(sig).merge(source.flatMap(a => Stream.eval_(sig.set(a))))
-    }
-
-  implicit def fs2SignalFunctor[F[_]](implicit F: Functor[F]): Functor[({type l[A]=Signal[F, A]})#l] =
-    new Functor[({type l[A]=Signal[F, A]})#l] {
-      def map[A, B](fa: Signal[F, A])(f: A => B): Signal[F, B] = fa.map(f)
     }
 }
