+++ core/src/main/scala/fs2/async/mutable/Queue.scala
@@ -10,7 +10,7 @@ import fs2.async.immutable
  * a queue may have a bound on its size, in which case enqueuing may
  * block until there is an offsetting dequeue.
  */
-trait Queue[F[_],A] { self =>
+trait Queue[F[_], A] { self =>
 
   /**
    * Enqueues one element in this `Queue`.
